Create Dual Axis Sun Tracking Solar Panel
Solar tracker, a system that positions an object at an angle relative to the sun. The most-common applications for solar trackers are positioning solar panels (photovoltaic panels) so that they remain perpendicular to the Sun’s rays and can determine the proper Sun’s direction. Solar trackers adjust the direction that a solar panel is facing according to the position of the Sun in the sky. By keeping the panel perpendicular to the Sun, more sunlight strikes the solar panel, less light is reflected, and more energy is absorbed. That energy can be converted into  power. Dual-axis trackers continually face the sun because they can move in two different directions. Types include tip-tilt and azimuth-altitude. Dual-axis tracking is typically used to orient a mirror and redirect sunlight along a fixed axis towards a stationary receiver. Because these trackers follow the sun vertically and horizontally they help obtain maximum solar energy generation at a time.

Working: Ldrs are used as the main light sensors. Two servo motors are fixed to the structure that holds the solar panel. LDRs sense the amount of sunlight falling on them. Four LDRs are divided into top, bottom, left and right.
For east – west tracking, the analog values from two top LDRs and two bottom LDRs are compared and if the top set of LDRs receive more light, the vertical servo will move in that direction and If the bottom LDRs receive more light, the servo moves in that direction.
For angular deflection of the solar panel, the analog values from two left LDRs and two right LDRs are compared. If the left set of LDRs receive more light than the right set, the horizontal servo will move in that direction and If the right set of LDRs receive more light, the servo moves in that direction.