Create Density Based Traffic Control System using ARIES board and IR Tranciever
The system contains IR sensors (transmitter and receiver) which will be mounted on the either side of the road on poles. It gets activated and receives the signal as the vehicles passes close by it. In this system, I will use IR sensors to measure the traffic density. We have to arrange one IR sensor for each road; these sensors always sense the traffic on that particular road. All these sensors are interfaced to the microcontroller. Based on these sensor outputs , controller detects the traffic and controls the traffic system.