Create Motorbike Helmet Detection
The YOLOv8 (You Only Look Once version 8) object detection technique was used to investigate
helmet identification in this research. Accurately
identifying helmets is essential to preventing accidents and ensuring compliance with safety laws
as safety concerns grow across a variety of industries, including motorcycling, sports, and construction sites. The cutting-edge real-time object identification model YOLOv8 is renowned for its accuracy and quickness. The study’s main objective is
to train and optimize the YOLOv8 model using a
unique dataset made up of pictures of people wearing helmets. Bounding boxes have been added to
the dataset to show where and how big the helmets
are in each image. With the help of these annotated
photographs, the YOLOv8 model is taught to recognize helmets in real-world situations by learning their features and attributes.